soulbind information done, need fix bd for covenant spells

+++ app/src/main/java/com/BlizzardArmory/model/warcraft/covenant/soulbind/TechTalentTree.kt
@@ -8,5 +8,5 @@ data class TechTalentTree(
 
     @SerializedName("key") val key: Key,
     @SerializedName("name") val name: String,
-    @SerializedName("id") val id: Int
+    @SerializedName("id") val id: Long
 )
